apply fixes for flaky stall tests
